interactive_correspondence/requirements.txt
flask==2.0.1
numpy==1.21.0
pillow==8.3.1
torch==1.9.0
torchvision==0.10.0 